requirements.txt
websockets==12.0
python-telegram-bot==20.7
aiohttp==3.9.1
asyncio-mqtt==0.16.1
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
websocket-client==1.7.0 